add save/load max ent fast

+++ src/mem/algo/__init__.py
@@ -1,7 +1,7 @@
 import numpy as np
 import numpy.typing as npt
 
-IdxType = np.int16
-FloatType = np.float32
+IdxType = np.int32
+FloatType = np.float64
 NDArrayInt = npt.NDArray[IdxType]
 NDArrayFloat = npt.NDArray[FloatType]
